requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-dev --output-file requirements.txt
altair==5.5.0
    # via streamlit
anyio==4.10.0
    # via httpx
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
greenlet==3.2.4 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via koderpad
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
narwhals==2.3.0
    # via altair
numpy==2.3.2
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.2
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.0
    # via streamlit
psycopg2-binary==2.9.10
    # via koderpad
pyarrow==21.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   koderpad
    #   streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via koderpad
streamlit==1.49.1
    # via
    #   koderpad
    #   streamlit-code-editor
streamlit-code-editor==0.1.22
    # via koderpad
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   referencing
    #   sqlalchemy
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via
    #   koderpad
    #   streamlit
